Move AOT test data to EmbedderTestContext
